完成数据统计接口

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -124,9 +124,9 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
     @Override
     public Result deleteById(Long id) {
 
-        int setmealCount = setmealService.count(new LambdaQueryWrapper<Setmeal>().eq(Setmeal::getCategoryId, id));
+        long setmealCount = setmealService.count(new LambdaQueryWrapper<Setmeal>().eq(Setmeal::getCategoryId, id));
 
-        int dishCount = dishService.count(new LambdaQueryWrapper<Dish>().eq(Dish::getCategoryId, id));
+        long dishCount = dishService.count(new LambdaQueryWrapper<Dish>().eq(Dish::getCategoryId, id));
 
         if (setmealCount > 0 || dishCount > 0){
             return Result.error("删除失败");
